backend/mail_config.py
from fastapi_mail import FastMail, ConnectionConfig
import os
from dotenv import load_dotenv

load_dotenv()

conf = ConnectionConfig(
    MAIL_USERNAME=os.getenv('USERMAIL'),
    MAIL_PASSWORD=os.getenv('PASSWORD'),
    MAIL_FROM=os.getenv('USERMAIL'),
    MAIL_PORT=587,
    MAIL_SERVER="smtp.gmail.com",
    MAIL_STARTTLS=True,
    MAIL_SSL_TLS=False,
    USE_CREDENTIALS=True,
    VALIDATE_CERTS=True
)

fm = FastMail(conf)