requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
sqlalchemy
python-dotenv
pydantic
fastapi-mail
bcrypt
python-jose[cryptography]
email-validator
